% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold-elementary.R
\name{manifold.torus}
\alias{manifold.torus}
\title{Flat Torus in Angle Coordinates}
\usage{
manifold.torus(d)
}
\arguments{
\item{d}{Number of circular factors.}
}
\value{
A manifold with d real angles, each interpreted modulo 2*pi.
}
\description{
Flat Torus in Angle Coordinates
}
\examples{
if (torch::torch_is_installed()) riem.random(manifold.torus(2))
}
